Update ch.py contact info and maintainers
If you want to list your email as well, feel free to submit a pull request or commit it yourself (if you have such permission(s).) Please only submit a pull to add yourself if you're an actual contributor or maintainer of ch.py. My list was based on https://github.com/Nullspeaker/ch.py/graphs/contributors as of June 18, 2014.